Updated Currency to a sealed class. Updated all ints to uints. We don't allow negative, so lets just not have them at all

+++ Backend/Console/Commands/Resolver/Pipelines/CurrencyUpdateResolver.cs
@@ -26,7 +26,7 @@ namespace Console.Commands.Resolver.Pipelines
             return new CurrencyUpdateArguments
             {
                 ActionType = actionType,
-                Amount = amount,
+                Amount = (uint) amount,
                 CurrencyType = currencyType
             };
         }
